bootstrap single page resume
Steps taken to make it look at current state
Inside the about section, give the h1 a display-1 class to make it stand out.
Give the p containing your "objective" a class of lead.
Set the margins of the hrs between the sections to 0 with the m-0 class.
Use the mb-5 class to add a 3rem bottom margin to each of the three divs containing the experience items.
Improve the spacing inside each experience item by applying mb-0 to the h3 containing the job title, and mb-3 to the company name subheading.
Using the margin settings that was implemented in the experience section, do the same thing to margins in the education section.
Add .5rem of padding to each of the three section elements, using the p-2 class.
In large viewports, it would look better to have more padding around the page and around each of the resume sections, but on small viewports we need to take advantage of all the space we have. To increase the padding to 3rem for large screens use a responsive padding utility class, p-lg-5, on each of the section elements. 
Prevent the h1 with the name from being obscured by the navbar by adding a pt-5 and m4-5 classes to the about section element.
Hide the large name in the about section when the navbar is collapsed by adding the d-none class to the h1 containing your name.
Show the large name in the about section when the viewport is md or larger by adding the d-md-block class to it.