reversed order of flex items
Completed following items to bring it to the current state. 
Added the d-flex and flex-column classes to each section elements.
To reverse the order of the flex items, change flex-column to flex-column-reverse in the experience section. The order of the experience items will be reversed. However, notice that the h2 elements are also converted to flex items and so they end up at the bottom of the list of items.
To make the header show up above the reversed flex items, remove it from the section element and place it just above the section. After doing this, apply pl-2 pl-lg-5 to it to maintain the same left margins.
To move the date information to a column on the right, you first need to make each of the section's subitems into a flex container. Do this by applying the d-flex and flex-column classes to the div surrounding each item.
Now that their parent div elements are flex containers, you can apply the text-md-right class to each of the div elements containing dates to move the date information to a column on the right for viewports that are md or larger.
Apply the text-primary class to each of the span elements containing the dates to make the text a light blue.
Add the flex-md-row class to the divs surrounding each item.
Create a new div that surrounds all of the blocks in an experience or education item except for the div containing the date information. This will create two flex items inside the container, such that the second one (with the dates) will appear to the right of the first.
Apply the text-nowrap class to the divs containing the dates, so that the dates do not wrap.